requirements.txt
phidata
openai
duckduckgo-search
python-dotenv
yfinance